Add --keep-files flag and force Drop of database in simulator to prevent issues with the Registry

+++ simulator/main.rs
@@ -136,7 +136,7 @@ fn testing_main(cli_opts: &SimulatorCLI) -> anyhow::Result<()> {
     println!("seed: {seed}");
     println!("path: {}", paths.base.display());
 
-    if result.is_ok() {
+    if !cli_opts.keep_files && result.is_ok() {
         paths.delete_all_files();
     }
